py: takes angles as floats not uint

+++ python/src/backward.cu
@@ -19,7 +19,7 @@ void backward_3d_cuda(
     nb::ndarray<curad::f32, nb::shape<3>, nb::device::cpu> vol_offset,
     nb::ndarray<curad::f32, nb::shape<nb::any, nb::any, nb::any>, nb::device::cuda, nb::c_contig>
         sinogram,
-    nb::ndarray<curad::u64, nb::shape<nb::any>, nb::device::cpu> angles,
+    nb::ndarray<curad::f32, nb::shape<nb::any>, nb::device::cpu> angles,
     nb::ndarray<curad::u64, nb::shape<2>, nb::device::cpu> det_shape,
     nb::ndarray<curad::f32, nb::shape<2>, nb::device::cpu> det_spacing,
     nb::ndarray<curad::f32, nb::shape<2>, nb::device::cpu> det_offset,
@@ -60,7 +60,7 @@ void backward_2d_cuda(
     nb::ndarray<curad::f32, nb::shape<2>, nb::device::cpu> vol_spacing,
     nb::ndarray<curad::f32, nb::shape<2>, nb::device::cpu> vol_offset,
     nb::ndarray<curad::f32, nb::shape<nb::any, nb::any>, nb::device::cuda, nb::c_contig> sino,
-    nb::ndarray<curad::u64, nb::shape<nb::any>, nb::device::cpu> angles, curad::u64 det_shape,
+    nb::ndarray<curad::f32, nb::shape<nb::any>, nb::device::cpu> angles, curad::u64 det_shape,
     curad::f32 det_spacing, curad::f32 det_offset, curad::f32 det_rotation, curad::f32 DSO,
     curad::f32 DSD, curad::f32 COR) {
 
@@ -78,5 +78,6 @@ void backward_2d_cuda(
     // TODO: make this call the same as forward_2d (both order and types of arguments)
     // TODO: Remove vol_extent, just compute it your-freaking-self :D
     curad::bp::backproject_2d(vol.data(), curad_vol_shape, curad_vol_spacing, curad_vol_offset,
-                              curad_vol_extent, sino.data(), det_shape, DSD, DSO, source, cpu_angles);
+                              curad_vol_extent, sino.data(), det_shape, DSD, DSO, source,
+                              cpu_angles);
 }
